🚧 Update mode on button release instead of press
This allows handing of a long press in future changes.

This requires button_state to be LOW at the start, which seems to be
working fine. There was prior reasoning for it being HIGH at the start,
which I'm having trouble understanding:

    Set it to high so then the shut down command inside handle_button runs on startup if the button is low

Also, on the very first press after uploading from PlatformIO, it
requires an extra press. Not sure why.